refactor: split benchmark reference dtos
